매치메이킹 서버와 동기화 작업 중 - 더미 테스트 2번 진행 중

+++ MasterServer/MasterServer/MasterServer.h
@@ -39,7 +39,7 @@ public:
 	// 사용자 정의 함수
 	//-----------------------------------------------------------
 	BattleServer*	FindBattleServerNo(int ServerNo);
-	CLIENT*			FindClientKey(int ClientKey);
+	CLIENT*			FindClientKey(UINT64 ClientKey);
 	bool			LanMonitorSendPacket(BYTE DataType);
 
 private:
